Rename convert file and make g4 repr traits accessible
