+++ src/geocel/g4/Convert.hh
@@ -3,7 +3,7 @@
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file geocel/g4/Convert.geant.hh
+//! \file geocel/g4/Convert.hh
 //---------------------------------------------------------------------------//
 #pragma once
