+++ src/geocel/g4/Repr.hh
@@ -1,0 +1,39 @@
+//----------------------------------*-C++-*----------------------------------//
+// Copyright 2024 UT-Battelle, LLC, and other Celeritas developers.
+// See the top-level COPYRIGHT file for details.
+// SPDX-License-Identifier: (Apache-2.0 OR MIT)
+//---------------------------------------------------------------------------//
+//! \file geocel/g4/Repr.hh
+//---------------------------------------------------------------------------//
+#pragma once
+
+#include <G4ThreeVector.hh>
+
+#include "corecel/io/Repr.hh"
+
+namespace celeritas
+{
+//---------------------------------------------------------------------------//
+template<>
+struct ReprTraits<G4ThreeVector>
+{
+    using value_type = std::decay_t<G4ThreeVector>;
+
+    static void print_type(std::ostream& os, char const* name = nullptr)
+    {
+        os << "G4ThreeVector";
+        if (name)
+        {
+            os << ' ' << name;
+        }
+    }
+    static void init(std::ostream& os) { ReprTraits<double>::init(os); }
+
+    static void print_value(std::ostream& os, G4ThreeVector const& vec)
+    {
+        os << '{' << vec[0] << ", " << vec[1] << ", " << vec[2] << '}';
+    }
+};
+
+//---------------------------------------------------------------------------//
+}  // namespace celeritas
